+++ .gitignore
@@ -1,0 +1,1 @@
+/leetcode_java_dcoder/out/
