+++ leetcode_cpp_dcoder/HOWTO_INSTALL.txt
@@ -1,0 +1,78 @@
+https://www.youtube.com/watch?v=DIw02CaEusY
+
+Install mingw-get is fine.
+
+Visual Studio Code
+CTRL SHIFT P
+
+C/Cpp: Edit Configurations…
+Creates c_cpp_properties.json
+Under intelliSenseMode add:
+
+           "browse": {
+                "path": [
+                    "${workspaceRoot}",
+                    "C:\\MinGW\\lib\\gcc\\mingw32\\6.3.0\\include\\c++"
+                ]
+            }
+
+
+CTRL SHIFT P
+Tasks: Configure Task
+Create tasks.json file from template
+Others
+
+   "version": "2.0.0",
+    "tasks": [
+        {
+            "label": "echo",
+            "type": "shell",
+            "command": "g++",
+            "args": [
+                "-g", "main.cpp"
+            ],
+            "group": {
+                "kind": "build",
+                "isDefault": true
+            }
+        }
+    ]
+
+CTRL SHIFT B - to build your program
+
+--------------
+
+Goto debug mode.
+Click GREEN ARROW
+C++ (GDB/LLDB)
+Creates a launch.json file
+
+   "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "(gdb) Launch",
+            "type": "cppdbg",
+            "request": "launch",
+            "program": "${workspaceFolder}/a.exe",
+            "args": [],
+            "stopAtEntry": false,
+            "cwd": "${workspaceFolder}",
+            "environment": [],
+            "externalConsole": true,
+            "MIMode": "gdb",
+            "miDebuggerPath": "C:\\MinGW\\bin\\gdb.exe",
+            "preLaunchTask": "echo",
+            "setupCommands": [
+                {
+                    "description": "Enable pretty-printing for gdb",
+                    "text": "-enable-pretty-printing",
+                    "ignoreFailures": true
+                }
+            ]
+        }
+    ]
+
+
+----------------------------
+
+
