+++ leetcode_java_dcoder/004_median_two_sorted_array.java
@@ -1,4 +1,4 @@
-package median_two_sorted_array;
+package _004_median_two_sorted_array;
 
 import java.text.DecimalFormat;
